+++ app/Http/Controllers/FormationController.php
@@ -8,9 +8,12 @@ use App\Concern\Invoice;
 use App\Concern\Tools;
 use App\Mail\ConfirmationInscriptionFormation;
 use App\Mail\SendInvoice;
+use App\Models\Evaluation;
+use App\Models\Evaluationstheme;
 use App\Models\Formation;
 use App\Models\Inscrit;
 use App\Models\Personne;
+use App\Models\Session;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Mail;
 
@@ -41,8 +44,24 @@ class FormationController extends Controller
         $user = session()->get('user');
         $personne = Personne::where('id', $user->id)->first();
         $inscriptions = [];
-        foreach ($personne->inscrits as $inscrit) {
+        foreach ($personne->inscrits->where('status', 1) as $inscrit) {
             $inscriptions[] = $inscrit->session_id;
+        }
+        foreach ($formation->sessions as $k => $session) {
+            $session_full = 0;
+            if ($session->ur_id != '') {
+                // on regarde si l'ur de la session est la même sque celui du user
+                if ($user->cartes[0]->urs_id != $session->ur_id) {
+                    $session_full = 1;
+                }
+            }
+            if ($session->club_id != '') {
+                // on regarde si l'ur de la session est la même sque celui du user
+                if ($user->cartes[0]->clubs_id != $session->club_id) {
+                    $session_full = 1;
+                }
+            }
+            $formation->sessions[$k]->full = $session_full;
         }
         $formation->location = strlen($formation->location) ? $formation->location : $this->getFormationCities($formation);
         return view('formations.detail', compact('formation', 'personne', 'inscriptions'));
@@ -53,7 +72,9 @@ class FormationController extends Controller
         $inscrit = Inscrit::where('monext_token', $request->token)->first();
         if ($inscrit) {
             $formation = $inscrit->session->formation;
-            $inscrit->delete();
+            if ($inscrit->secure_code != '') {
+                $inscrit->delete();
+            }
             return redirect()->route('formations.detail', $formation->id)->with('error', "Votre paiement a été annulé");
         } else {
             return redirect()->route('formations.accueil')->with('error', "Le paiement a été annulé");
@@ -67,7 +88,7 @@ class FormationController extends Controller
             $inscrit = Inscrit::where('monext_token', $request->token)->where('attente_paiement', 1)->first();
             if ($inscrit) {
                 // on met à jour le flag attente_paiement à 0 pour l'inscrit
-                $data = ['attente_paiement' => 0, 'status' => 1];
+                $data = ['attente_paiement' => 0, 'status' => 1, 'secure_code' => null];
                 $inscrit->update($data);
                 $formation = $inscrit->session->formation;
 
@@ -99,6 +120,41 @@ class FormationController extends Controller
     public function attentePaiementValidation($formation_id)
     {
         return redirect()->route('formations.detail', $formation_id)->with('success', "Si vous avez procédé au paiement par virement de votre inscription, celle-ci sera traitée d'ici quelques minutes et un email vous informera de sa prise en compte");
+    }
+
+    public function payWithSecureCode($secure_code) {
+        $user = session()->get('user');
+        $inscrit = Inscrit::where('secure_code', $secure_code)->first();
+        if (!$inscrit) {
+            return redirect()->route('accueil');
+        }
+        if ($inscrit->personne_id != $user->id) {
+            return redirect()->route('accueil');
+        }
+
+        return view('formations.paiement', compact('inscrit'));
+    }
+
+    public function evaluation($md5) {
+        $session = Session::whereRaw("md5(id) = '$md5'")->first();
+        if (!$session) {
+            return redirect()->route('accueil')->with('error', "Le lien d'évaluation est désormais invalide");
+        }
+
+        // on regarde si une évlautaion a déjà été faite pour cette session
+        $user = session()->get('user');
+        $evaluation = Evaluation::where('session_id', $session->id)->where('personne_id', $user->id)->first();
+        if ($evaluation) {
+            return redirect()->route('accueil')->with('error', "Vous avez déjà évalué cette formation");
+        }
+
+        // on cherche les éléments de validation
+        $themes = Evaluationstheme::orderBy('position')->get();
+//        foreach ($themes as $theme) {
+//            dd($theme->evaluationsitems->sortBy('position'));
+//        }
+
+        return view('formations.evaluation', compact('session', 'user', 'themes'));
     }
 
 }
